+++ lib/src/types/gen/cosmos/consensus/v1/tx.pbserver.dart
@@ -1,0 +1,38 @@
+///
+//  Generated code. Do not modify.
+//  source: cosmos/consensus/v1/tx.proto
+//
+// @dart = 2.12
+// ignore_for_file: annotate_overrides,camel_case_types,constant_identifier_names,deprecated_member_use_from_same_package,directives_ordering,library_prefixes,non_constant_identifier_names,prefer_final_fields,return_of_invalid_type,unnecessary_const,unnecessary_import,unnecessary_this,unused_import,unused_shown_name
+
+import 'dart:async' as $async;
+
+import 'package:protobuf/protobuf.dart' as $pb;
+
+import 'dart:core' as $core;
+import 'tx.pb.dart' as $32;
+import 'tx.pbjson.dart';
+
+export 'tx.pb.dart';
+
+abstract class MsgServiceBase extends $pb.GeneratedService {
+  $async.Future<$32.MsgUpdateParamsResponse> updateParams($pb.ServerContext ctx, $32.MsgUpdateParams request);
+
+  $pb.GeneratedMessage createRequest($core.String method) {
+    switch (method) {
+      case 'UpdateParams': return $32.MsgUpdateParams();
+      default: throw $core.ArgumentError('Unknown method: $method');
+    }
+  }
+
+  $async.Future<$pb.GeneratedMessage> handleCall($pb.ServerContext ctx, $core.String method, $pb.GeneratedMessage request) {
+    switch (method) {
+      case 'UpdateParams': return this.updateParams(ctx, request as $32.MsgUpdateParams);
+      default: throw $core.ArgumentError('Unknown method: $method');
+    }
+  }
+
+  $core.Map<$core.String, $core.dynamic> get $json => MsgServiceBase$json;
+  $core.Map<$core.String, $core.Map<$core.String, $core.dynamic>> get $messageJson => MsgServiceBase$messageJson;
+}
+