+++ lib/src/types/types.dart
@@ -3,6 +3,7 @@ import 'package:motor_flutter/motor_flutter.dart';
 export 'gen/motor/v1/request.pb.dart';
 export 'gen/motor/v1/response.pb.dart';
 export 'gen/schema/v1/schema.pb.dart';
+export 'gen/schema/v1/document.pb.dart';
 export 'gen/registry/did.pb.dart';
 export 'gen/service/v1/discover.pb.dart';
 export 'gen/registry/tx.pb.dart';
@@ -32,6 +33,22 @@ extension CreateAccountWithKeysResponseExt on CreateAccountWithKeysResponse {
       address: address,
       did: whoIs.didDocument.id,
       password: pwd,
+    );
+  }
+}
+
+extension SchemaDefinitionExt on SchemaDefinition {
+  SchemaDocument getDefaultDocument() {
+    return SchemaDocument(
+      creator: creator,
+      did: did,
+      definition: this,
+      fields: List<SchemaDocumentValue>.generate(fields.length, (index) {
+        return SchemaDocumentValue(
+          name: fields[index].name,
+          field_2: fields[index].field_2,
+        );
+      }),
     );
   }
 }